Fixed issue ET-34.
Bug was with run finding and sorting. The sorting would try to
compare the completed attribute of a run which may not exist in
the event of an incomplete run. Comparing a datetime and a None
would result in an exception. This occurs before any processing
or digestion of a run and is not caught by the exception wrapper
around run ingestion.