Adds the ability to move beyond the five default forks.
The number of forks used when running openstack-ansible snitch.yml
can greatly affect the run time. For large environments, 5 forks
may not be enough to finish within a reasonable amount of time.

This change allows one to specify the number of forks. If this is
not specified, the number of forks will default to:
    max(number of threads - 2, 8)